is a string unique characters
first, need to consider if string is limited to ASCII or unicode. If constrained to ASCII, could check length of String if more than 128 length it has to have repeating characters and you wouldnt need to compare anything... I used a HashSet, compare length of set to string. This is O(n) however because have to go through entire string to add to hashset. A more efficient way to determine uniqueness would be to do HashTable, if HashTable(string[i]) already exists, no need to continue while worst case is still O(n) would not always be O(n)

if you cannot use additional data structures, use nested forLoop. This is O(n2) but if memory constraints, would need to do this more timely way.