PermutationPalindrome
checks if  permutation of a string, omiting blank spaces, could be a palindrome. Push string characters to an array. Then adds character to a frequency table/hashMap. Then visit every key checking if value is odd or even. Regardless of the amount of permuations, a palindrome can only have one or none odd occuring characters. Remeber, when checking chars use "==" when checking a String use ".equals()"!